test(procmgr): rename catalog_list e2e module to list
Aligns with describe.rs and other command-named e2e modules.
